add can't delete user if have cart fix some css

+++ views/v_admin_client_edit.php
@@ -9,9 +9,9 @@ if (@isset($_POST['btn_update'])) {
         $error['fullname'] = "Không được để trống Họ Và Tên";
     }
     foreach (getUser() as $item) {
-        if($item['username'] == $_POST['username']) {
+        if ($item['username'] == $_POST['username']) {
             $error['username'] = "Tên đăng nhập này đã được sử dụng";
-        }elseif($item['email'] == $_POST['email']) {
+        } elseif ($item['email'] == $_POST['email']) {
             $error['email'] = "Email này đã được sử dụng";
         }
     }
@@ -112,13 +112,27 @@ if (@isset($_POST['btn_update'])) {
 if (@isset($_POST['btn_delete'])) {
     $id = $_GET['id'];
     //Thư mục chứa file Delete
-    $delete_dir = './upload/users/';
-    //Đường dẫn của file sau khi Delete
-    $delete_file = $delete_dir . $userInfo[0]['img'];
-    echo $delete_file;
-    deleteUser($id);
-    unlink($delete_file);
-    header("Location: ?mod=admin&act=client");
+    $getAllBilById = get_allBill($id);
+    foreach ($getAllBilById as $item) {
+        if ($item['status'] == 1 || $item['status'] == 2 || $item['status'] == 3) {
+            $error['status'] = "Không thể xóa User này vì hiện đang có đơn hàng đang chuẩn bị hoặc đang giao";
+        } else {
+
+        }
+    }
+
+    if (isset($error['status']) && $error['status']) {
+
+    } else {
+        $delete_dir = './upload/users/';
+        //Đường dẫn của file sau khi Delete
+        $delete_file = $delete_dir . $userInfo[0]['img'];
+        echo $delete_file;
+        deleteUser($id);
+        unlink($delete_file);
+        header("Location: ?mod=admin&act=client");
+    }
+
 
 }
 if (@isset($_POST['btn_cancelled'])) {
@@ -151,8 +165,7 @@ if (@isset($_POST['btn_cancelled'])) {
                         <li>
                             <div class="col-12 d-flex">
                                 <div class="col-2">
-                                    <img class="notifiAdminImg"
-                                        src="./upload/users/<?php echo $getUser['img'] ?>" alt="">
+                                    <img class="notifiAdminImg" src="./upload/users/<?php echo $getUser['img'] ?>" alt="">
                                 </div>
                                 <div class="col-10">
                                     <p class="notifiAdminText body-small"><strong>
@@ -205,8 +218,7 @@ if (@isset($_POST['btn_cancelled'])) {
                 $getID = $_SESSION['admin']['id_user'];
                 $getUser = getUserById($getID);
                 ?>
-                <img style="" class="btnShowFeature"
-                    src="./upload/users/<?php echo $getUser['img'] ?>" alt="">
+                <img style="" class="btnShowFeature" src="./upload/users/<?php echo $getUser['img'] ?>" alt="">
                 <ul class="showFeatureAdminHeader box-shadow1">
 
                     <li><a class="body-small" href="#statisticalChart">Thống kê đơn hàng</a></li>
@@ -242,7 +254,7 @@ if (@isset($_POST['btn_cancelled'])) {
                 </div>
                 <form enctype="multipart/form-data" action="" method="POST">
                     <div class="sliderDashboard_order-add-create sliderDashboard_order-detail rounded-4">
-                        <div class="body_sliderDashboard_order-add-create p20 row">
+                        <div class="body_sliderDashboard_order-add-create p20 row col-12">
                             <div class="col-7">
                                 <div class="left-order-add-create">
                                     <h2>Họ Và Tên</h2>
@@ -251,7 +263,7 @@ if (@isset($_POST['btn_cancelled'])) {
                                         aria-label="default input example">
                                     <?php
                                     if (isset($error['fullname']) && !empty($error['fullname']))
-                                        echo "<p class='text-danger text-error'>{$error['fullname']}</p>";
+                                        echo "<p class='text-danger text-error title-medium'>{$error['fullname']}</p>";
                                     ?>
                                 </div>
                                 <div class="left-order-add-create">
@@ -261,7 +273,7 @@ if (@isset($_POST['btn_cancelled'])) {
                                         aria-label="default input example">
                                     <?php
                                     if (isset($error['username']) && !empty($error['username']))
-                                        echo "<p class='text-danger text-error'>{$error['username']}</p>";
+                                        echo "<p class='text-danger text-error title-medium'>{$error['username']}</p>";
                                     ?>
                                 </div>
 
@@ -272,7 +284,7 @@ if (@isset($_POST['btn_cancelled'])) {
                                         aria-label="default input example">
                                     <?php
                                     if (isset($error['password']) && !empty($error['password']))
-                                        echo "<p class='text-danger text-error'>{$error['password']}</p>";
+                                        echo "<p class='text-danger text-error title-medium'>{$error['password']}</p>";
                                     ?>
                                 </div>
                                 <div class="left-order-add-create">
@@ -281,7 +293,7 @@ if (@isset($_POST['btn_cancelled'])) {
                                         placeholder="Email" aria-label="default input example">
                                     <?php
                                     if (isset($error['email']) && !empty($error['email']))
-                                        echo "<p class='text-danger text-error'>{$error['email']}</p>";
+                                        echo "<p class='text-danger text-error title-medium'>{$error['email']}</p>";
                                     ?>
                                 </div>
                                 <div class="left-order-add-create">
@@ -290,7 +302,7 @@ if (@isset($_POST['btn_cancelled'])) {
                                         placeholder="Nhập Số Điện Thoại" aria-label="default input example">
                                     <?php
                                     if (isset($error['phone']) && !empty($error['phone']))
-                                        echo "<p class='text-danger text-error'>{$error['phone']}</p>";
+                                        echo "<p class='text-danger text-error title-medium'>{$error['phone']}</p>";
                                     ?>
                                 </div>
                                 <div class="describe-order_detail">
@@ -299,7 +311,7 @@ if (@isset($_POST['btn_cancelled'])) {
                                         value="<?php echo $userInfo[0]['bio'] ?>"><?php echo $userInfo[0]['bio'] ?></textarea>
                                     <?php
                                     if (isset($error['bio']) && !empty($error['bio']))
-                                        echo "<p class='text-danger text-error'>{$error['bio']}</p>";
+                                        echo "<p class='text-danger text-error title-medium'>{$error['bio']}</p>";
                                     ?>
                                 </div>
 
@@ -331,8 +343,7 @@ if (@isset($_POST['btn_cancelled'])) {
                                             <?php
                                         } else {
                                             ?>
-                                            <td><img
-                                                    src="./upload/users/<?php echo $userInfo[0]['img'] ?>">
+                                            <td><img src="./upload/users/<?php echo $userInfo[0]['img'] ?>">
                                             </td>
                                             <?php
                                         }
@@ -359,6 +370,10 @@ if (@isset($_POST['btn_cancelled'])) {
                                                 class="btn_cancelled">
                                         </div>
                                     </div>
+                                    <?php
+                                    if (isset($error['status']) && !empty($error['status']))
+                                        echo "<p class='text-danger text-error title-medium'>{$error['status']}</p>";
+                                    ?>
                                 </div>
                             </div>
                         </div>
